Email-Context/email_gateway/requirements.txt
aiodataloader==0.4.0
aniso8601==9.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dependency-injector==4.41.0
Flask==2.3.3
Flask-Mail==0.9.1
gevent==23.9.1
gevent-websocket==0.10.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphql-server==3.0.0b7
greenlet==3.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
promise==2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.31.0
ruff==0.1.14
Rx==1.6.3
six==1.16.0
starkbank-ecdsa==2.2.0
typing_extensions==4.8.0
urllib3==2.0.7
watchdog==3.0.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
